requirements.txt
fastapi==0.121.0
uvicorn==0.38.0

SQLAlchemy==2.0.44
sqlmodel==0.0.18
psycopg==3.2.12
psycopg2-binary==3.2.11
aiosqlite==0.21.0

python-dotenv==1.2.1
pydantic==2.11.9

passlib==1.7.4
python-jose==3.5.0
bcrypt==4.0.1

langchain==1.0.3
langchain-google-genai==3.0.1
langgraph==1.0.2

httpx==0.28.1
requests==2.32.5

streamlit==1.28.1
pandas==2.1.3
plotly==5.18.0